requirements.txt
streamlit>=1.28.0
fpdf>=1.7.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.24.0
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.2
ollama>=0.1.7